Following bugs resolved:
Playing screen : Click on the Add Playlist option, and without selecting any playlist, click on the Add button — an error appears.
Select song : When a song is already added to the queue, an incorrect toast message is displayed.
Playing screen : When we add music to any playlist, the success toast message is not displayed. Also, if the music is already added to the playlist, the notice toast message is not displayed.